Revert "Remove cron verification tasks"
This reverts commit ba3a5a2d469d0d91dd62abbb7ccf05aec50d82d6.

It turns out that these tests were legitimately failing because
the log directory specified in the cron jobs was inaccessible by
the user who runs the cron jobs.

That was fixed in commit 51d6fddfe0aa38b942cc01a2ed7af2f837b1a44d and
so the tests can and should be added back.

Additionally, when the cron jobs fail to run it could result in
catastrophic fires for whoever is relying on this role and so we should
make best efforts to ensure that they do run.  Which is why is it a good
idea to keep these tests.